requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Hyperparameter optimization
optuna>=3.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration management
pyyaml>=6.0

# Data processing
scipy>=1.11.0

# Optional: for better progress bars
tqdm>=4.65.0

# Optional: for model interpretation
shap>=0.42.0
